Se genera planilla por pago de actividad economica

+++ database/migrations/2020_01_03_103243_create_bank_rates_table.php
@@ -15,7 +15,7 @@ class CreateBankRatesTable extends Migration
     {
         Schema::create('bank_rates', function (Blueprint $table) {
             $table->increments('id');
-            $table->float('value_rate');
+            $table->decimal('value_rate',11,3);
             $table->timestamps();
         });
     }
